refactoring: migrate actions
