+++ .env.example
@@ -26,7 +26,6 @@ MIX_PANEL_TOKEN=
 # Treasuery
 CHOMP_TREASURY_PRIVATE_KEY=
 CHOMP_SQUADS=
-NEXT_PUBLIC_TREASURY_PUBLIC_ADDRESS=
 
 # Chomp Bot [Optional]
 BOT_API_KEY=jFk207-3eghi6-ql0cvp-u9wx45
